+++ app/client/authentication/check_hostname.js
@@ -1,0 +1,8 @@
+const os = require('node:os');
+
+export async function check_hostname() {
+
+    machine_hostname = os.hostname()
+    
+    return
+}