+++ app/client/authentication/check_mac.js
@@ -1,0 +1,8 @@
+const os = require('node:os');
+
+export async function check_mac() {
+
+    mac = os.networkInterfaces().en0[0].mac
+
+    return
+}